Create pottyTrainingAppPage
This is a webpage page made with HTML, CSS, and Bootstrap. It is a page that uses clever, and modern, techniques for advertising an app that would be available through the iPhone App Store. 

I will be implementing SQL databases as well as PHP to modify them.

Note, the actual app does not yet exist, or is not finished.